the-big-fan/python/requirements.txt
-e .
aws-cdk.core==1.83.0
aws-cdk.aws_apigateway==1.83.0
aws-cdk.aws-lambda==1.83.0
aws-cdk.aws-lambda-event-sources==1.83.0
aws-cdk.aws-iam==1.83.0
aws-cdk.aws-sns==1.83.0
aws-cdk.aws-sns-subscriptions==1.83.0
aws-cdk.aws-sqs==1.83.0
